refactor(signer): pull tests from old repo & remove dependency on parent instance
In the old repo "FoxySigner" was a field of a FoxyAPI instance, and
therefore had access to its clientSecret. It is now an independent
utility. The first step to make this code adherent to the new codebase
was to remove this dependency. This was only relevant in the test file
because the class itself was already receiving clientSecret as a
parameter.

There is still two issues, though.
1- tests are dependent on the existence of certain html files on
the filesystem in order to test the ability to sign files. This
dependency must be removed from the tests.
2- Signer should become a simple function, as happened with the other
utilities from the FoxyAPI in the previous repository.